pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eplacement-kyc-verifier"
version = "0.1.0"
description = "Local KYC evidence verification sidecar for the ePlacement algorithm."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "streamlit>=1.42.0",
  "pandas>=2.2.0",
  "openpyxl>=3.1.0",
  "PyYAML>=6.0.1",
  "requests>=2.32.0",
  "pydantic>=2.8.0",
  "python-multipart>=0.0.9",
  "PyMuPDF>=1.24.0",
  "Pillow>=10.4.0",
  "pytesseract>=0.3.10",
  "ocrmypdf>=16.4.0",
  "langflow>=1.0.19",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0"
]
paddle = [
  "paddleocr>=2.8.1"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
